+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 ### Breaking Changes
 * :x: ``dimension`` module moved to ``soccer.dimensions``.
 * :x: ``spot_scale`` argument of ``Pitch`` classes default changed \
-to 0.04 from 0.002. This produces approximately the same size center \
+to 0.024 from 0.002. This produces approximately the same size center \
 spots as previously. The spot scale is now set in relation to the \
-center circle, i.e. 4% of the center circle diameter, rather than the \
-previous length of the pitch.
+center circle, i.e. 2.4% of the center circle diameter, rather than in \
+relation to the length of the pitch as previously.
 
 :rocket: Version 1.5.1
 ----------------------