+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+:rocket: Version 1.6.0
+----------------------
+
+### Breaking Changes
+* :x: ``dimension`` module moved to ``soccer.dimensions``.
+* :x: ``spot_scale`` argument of ``Pitch`` classes default changed \
+to 0.024 from 0.002. This produces approximately the same size center \
+spots as previously. The spot scale is now set in relation to the \
+center circle, i.e. 2.4% of the center circle diameter, rather than in \
+relation to the length of the pitch as previously.
+
 :rocket: Version 1.5.1
 ----------------------